+++ db/migrate/20210128114747_add_supplier_id_to_order.rb
@@ -1,0 +1,5 @@
+class AddSupplierIdToOrder < ActiveRecord::Migration[6.0]
+  def change
+    add_reference :orders, :supplier, foreign_key: true
+  end
+end
